refactor(ApiHubAPI): splitting portal and registry types - SWG-16338
