+++ src/api-hub/client/index.ts
@@ -3,5 +3,6 @@ export {
   ApiHubConfiguration,
   type ApiHubConfigurationParameters,
 } from "./configuration.js";
+export * from "./portal-types.js";
+export * from "./registry-types.js";
 export { type ApiHubToolParams, TOOLS } from "./tools.js";
-export * from "./types.js";
